Working again with permafrost dict output.
